fix(rrweb): WIP: refactor sideeffects from rrweb/record
